+++ cfgov/unprocessed/js/routes/on-demand/expandable.js
@@ -1,5 +1,5 @@
 /* ==========================================================================
-   Scripts for /browse-basic/.
+   Scripts for Expandable Molecule.
    ========================================================================== */
 
 'use strict';
@@ -9,7 +9,6 @@ require( '../../modules/polyfill/query-selector' );
 
 // List of organisms used.
 var Expandable = require( '../../molecules/Expandable' );
-var ExpandableGroup = require( '../../organisms/ExpandableGroup' );
 
 var selector = '.m-expandable';
 var expandables = document.querySelectorAll( selector );
@@ -19,11 +18,3 @@ for ( var i = 0, len = expandables.length; i < len; i++ ) {
   expandable = new Expandable( expandables[i] );
   expandable.init();
 }
-
-selector = '.o-expandable-group';
-var expandableGroup = document.querySelectorAll( selector );
-var regularExpandableGroup = new ExpandableGroup( expandableGroup[0] );
-var accordionExpandableGroup = new ExpandableGroup( expandableGroup[1] );
-
-regularExpandableGroup.init();
-accordionExpandableGroup.init();